// src/app/cc-date-time-picker/helper/enums.ts
export enum SelectMode {
  Range = 'range',
  Single = 'single'
}

export enum PickerType {
  DateTime = 'datetime',
  Date = 'date',
  Time = 'time',
  TimeCustom = 'timecustom',
  Month = 'month',
  Year = 'year',
  Week = 'week'
}

export enum PickerTriggerType {
  Button = 'button',
  Link = 'link'
}

export enum DateType {
  From = 1,
  To = 2
}

export enum MomentFormat {
  Auto = 'auto',
  DateTime = 'L LTS',
  Date = 'L',
  Year = 'YYYY',
  Month = 'MMMM YYYY',
  Week = 'L',
  Day = 'L',
  Time = 'LTS'
}